:sparkles: added base android auth
