+++ android/settings.gradle
@@ -1,4 +1,6 @@
 rootProject.name = 'RockNative'
+include ':react-native-touch-id-android'
+project(':react-native-touch-id-android').projectDir = new File(rootProject.projectDir, '../node_modules/react-native-touch-id-android/android')
 include ':react-native-device-info'
 project(':react-native-device-info').projectDir = new File(rootProject.projectDir, '../node_modules/react-native-device-info/android')
